merge Serializer + Deserializer into Stream
We cannot do two mutable borrows on the same underlying stream to add both a
Serializer and Deserializer wrapper. Instead, we implement only one wrapper
(named Stream). Depending on whether the underlying stream can Read or Write
(or both), appropriate sets of helper methods in the wrapper get enabled or
disabled - and we borrow only once.